+++ src/components/common/index.ts
@@ -3,3 +3,4 @@ export * from "./Text";
 export * from "./RecordCard";
 export * from "./Counter";
 export * from "./ImageBanner";
+export * from "./QuestionButton";
