+++ src/styles/colors.ts
@@ -1,7 +1,7 @@
 export const colors = {
   background: "#000000",
   card: "#333333",
-  cardHover: "#747474",
+  qnaClicked: "#747474",
   qnaCard: "#1E1E1E",
   text: "#eeeeee",
 };
